+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.3 (28 August 2024)
+## v0.3 (29 August 2024)
 
 - Extra game details are now pulled from the `/games/{game_id}/platforms/{platform_id}`
   endpoint after the initial details have been pulled from
@@ -19,6 +19,11 @@
 - Major restructuring so multiple files are exported for different sets of data. This is
   to work with the different data shapes, as well as the Microsoft Access limit of 255
   columns, and line lengths of 65,534 characters.
+
+- Because Microsoft Access is terrible with date formats, let alone partial dates, a
+  `releases_release_year` column is added next to existing `releases_release_date`
+  columns. This only lists the years, so comparison queries can be easily run against
+  them as integers.
 
 - A bug has been fixed where the rate limit was being interpreted as a string instead of
   an integer.